refactor: simplify dao contract hooks and add clientId to proposals
This commit refactors the DAO contract hooks to make them cleaner and more maintainable. More importantly, it also adds a clientId field to the proposals. The clientId will help better integrate our chain with other systems. The getContractAddress() function was removed, and a contractAddress constant was used instead. The use of an environment file was also introduced for some global constants.