refactor: update proposal screen and remove proposal action form component
Refactored the code in proposal-screen.js to integrate voting and comment functionality directly into the proposal page without the use of the proposal-action-form component. This was achieved through modifications to the 'useFeedItems' and 'handleFormSubmit' functions, and the addition of new import statements. Removed the entire 'ProposalActionForm' component as the functionalities it provided are now handled directly in the proposal screen.