fix: comment out most functionalities in farcaster proposal casts route
The majority of the functions in the file farcaster-proposal-casts/route.js have been commented out, effectively disabling them. This includes both imports and key functionalities such as createCanonicalProposalUrl and fetchProposalCasts, as well as several conditions in the POST function. Whatever the reason for this broad-sweeping change, developers should be aware that farcaster-proposal-casts/route.js is currently not operational in terms of its original scope and functionalities.