fix: disable fetching of proposal and candidate feedback
This commit disables the fetching of proposal and candidate feedbacks in the 'fetchVoterActivity' function within 'store.js' file. This change affects the way proposalIds and candidateIds arrays are populated, as they no longer include feedback-related entries. More underlying implications of this change need to be studied as part of further work.