feat: add farcaster accounts route
Added a new route for Farcaster accounts. This route is meant to return Farcaster accounts which are matching a verified Ethereum account address. This can be useful in cases where the system needs to fetch accounts associated with a certain Ethereum address, validate the Ethereum address, and check the existence of keys for each account.