refactor: temporary deactivate unused queries for proposal candidates and feedbacks
Remove the execution of the fetch operation for candidate ids and feedbacks. The purpose of this commit is to halt unused queries in the code to improve execution speed and optimize the code. Ideally, these parts will either be removed completely or updated, depending on the future needs of the application.