feat(browse-screen): add vote overview dialog
The update introduces a new `proposalVotesDialog` component to the `browse-screen` component. This feature enhancement is aimed to provide a detailed overview of votes to the users, aiding in their decision-making process. A link has also been established from the votes tag group to the vote overview for ease of access.